replace symlinks with actual files to prevent Windows CLI issues when building from source.
